requirements.txt
fastapi
langchain
langchain-community
python-decouple
Pillow